# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PDAM Tiara Material Tracking & Verification System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trackings:
    get:
      operationId: listTrackings
      tags: [trackings]
      summary: List trackings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: string
        - name: isOverdue
          in: query
          schema:
            type: boolean
        - name: assignedTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of trackings with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingListResponse"
    post:
      operationId: createTracking
      tags: [trackings]
      summary: Issue material from warehouse and create tracking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingInput"
      responses:
        "201":
          description: Tracking created with QR code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tracking"
  /trackings/{id}:
    get:
      operationId: getTracking
      tags: [trackings]
      summary: Get tracking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingDetail"
  /trackings/by-code/{code}:
    get:
      operationId: getTrackingByCode
      tags: [trackings]
      summary: Get tracking by tracking code (for QR scan)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingDetail"
  /trackings/{id}/receive:
    post:
      operationId: receiveTracking
      tags: [trackings]
      summary: Confirm branch reception by scanning QR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveInput"
      responses:
        "200":
          description: Reception confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tracking"
  /trackings/{id}/proof:
    post:
      operationId: submitInstallationProof
      tags: [trackings]
      summary: Submit installation proof (photo + GPS)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallationProofInput"
      responses:
        "201":
          description: Proof submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallationProof"
    get:
      operationId: getInstallationProof
      tags: [trackings]
      summary: Get installation proof for a tracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Installation proof
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallationProof"
  /trackings/{id}/events:
    get:
      operationId: getTrackingEvents
      tags: [trackings]
      summary: Get event timeline for a tracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracking events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackingEvent"
  /trackings/{id}/flag:
    post:
      operationId: flagTracking
      tags: [trackings]
      summary: Flag tracking for audit review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagInput"
      responses:
        "200":
          description: Tracking flagged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trackings/{id}/qr:
    get:
      operationId: getQrCode
      tags: [trackings]
      summary: Get QR code image for a tracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCodeResponse"

  # Dashboard
...
  /proofs/{id}/review:
    post:
      operationId: reviewProof
      tags: [dashboard]
      summary: SPI reviews a flagged proof (approve/reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProofReviewInput"
      responses:
        "200":
          description: Proof reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallationProof"

  # Reports
...
  /upload/photo:
    post:
      operationId: uploadPhoto
      tags: [trackings]
      summary: Upload installation photo (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoUploadInput"
      responses:
        "200":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUploadResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth
...
    Tracking:
      type: object
      required: [id, trackingCode, qtyIssued, issuedAt, slaDeadline, status, createdAt]
      properties:
        id:
          type: string
        trackingCode:
          type: string
        materialRequestId:
          type: ["string", "null"]
        materialName:
          type: ["string", "null"]
        materialCode:
          type: ["string", "null"]
        branchName:
          type: ["string", "null"]
        branchId:
          type: ["string", "null"]
        qtyIssued:
          type: integer
        issuedById:
          type: ["string", "null"]
        issuedByName:
          type: ["string", "null"]
        issuedAt:
          type: string
          format: date-time
        slaDeadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [dikirim, diterima_cabang, menunggu_verifikasi, terverifikasi, kritis]
        qrCodeUrl:
          type: ["string", "null"]
        riskScore:
          type: ["number", "null"]
        hoursRemaining:
          type: ["number", "null"]
        isOverdue:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TrackingDetail:
      type: object
      required: [id, trackingCode, qtyIssued, issuedAt, slaDeadline, status, createdAt, events]
      properties:
        id:
          type: string
        trackingCode:
          type: string
        materialRequestId:
          type: ["string", "null"]
        materialName:
          type: ["string", "null"]
        materialCode:
          type: ["string", "null"]
        materialUnit:
          type: ["string", "null"]
        branchName:
          type: ["string", "null"]
        branchId:
          type: ["string", "null"]
        qtyIssued:
          type: integer
        issuedById:
          type: ["string", "null"]
        issuedByName:
          type: ["string", "null"]
        issuedAt:
          type: string
          format: date-time
        slaDeadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [dikirim, diterima_cabang, menunggu_verifikasi, terverifikasi, kritis]
        qrCodeUrl:
          type: ["string", "null"]
        riskScore:
          type: ["number", "null"]
        hoursRemaining:
          type: ["number", "null"]
        isOverdue:
          type: boolean
        createdAt:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
        proof:
          $ref: "#/components/schemas/InstallationProof"
    TrackingListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Tracking"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TrackingInput:
      type: object
      required: [materialRequestId, qtyIssued, slaDays]
      properties:
        materialRequestId:
          type: string
        qtyIssued:
          type: integer
        slaDays:
          type: integer
        notes:
          type: ["string", "null"]
    ReceiveInput:
      type: object
      required: [qrScannedCode]
      properties:
        qrScannedCode:
          type: string
        qtyReceived:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]

    # Tracking Events
    TrackingEvent:
      type: object
      required: [id, trackingId, step, occurredAt]
      properties:
        id:
          type: string
        trackingId:
          type: string
        step:
          type: string
          enum: [keluar_gudang, diterima_cabang, dipasang, selesai]
        actorId:
          type: ["string", "null"]
        actorName:
          type: ["string", "null"]
        gpsLat:
          type: ["number", "null"]
        gpsLng:
          type: ["number", "null"]
        photoUrl:
          type: ["string", "null"]
        qrScannedCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        occurredAt:
          type: string
          format: date-time

    # Installation Proof
    InstallationProof:
      type: object
      required: [id, trackingId, qtyUsed, qtyReturned, photoUrl, photoTakenAt, gpsLat, gpsLng, submittedAt]
      properties:
        id:
          type: string
        trackingId:
          type: string
        submittedById:
          type: ["string", "null"]
        submittedByName:
          type: ["string", "null"]
        qtyUsed:
          type: integer
        qtyReturned:
          type: integer
        photoUrl:
          type: string
        photoTakenAt:
          type: string
          format: date-time
        gpsLat:
          type: number
        gpsLng:
          type: number
        gpsAccuracyMeters:
          type: ["number", "null"]
        gpsProvider:
          type: ["string", "null"]
        isMockLocation:
          type: boolean
        withinServiceArea:
          type: ["boolean", "null"]
        flaggedForReview:
          type: boolean
        reviewStatus:
          type: ["string", "null"]
          enum: [pending, approved, rejected, null]
        reviewedByName:
          type: ["string", "null"]
        flagReasons:
          type: array
          items:
            type: string
        submittedAt:
          type: string
          format: date-time
    InstallationProofInput:
      type: object
      required: [qtyUsed, qtyReturned, photoUrl, photoTakenAt, gpsLat, gpsLng]
      properties:
        qtyUsed:
          type: integer
        qtyReturned:
          type: integer
        photoUrl:
          type: string
        photoTakenAt:
          type: string
          format: date-time
        gpsLat:
          type: number
        gpsLng:
          type: number
        gpsAccuracyMeters:
          type: ["number", "null"]
        gpsProvider:
          type: ["string", "null"]
        isMockLocation:
          type: boolean
        deviceFingerprint:
          type: ["string", "null"]
    FlagInput:
      type: object
      properties:
        reason:
          type: string
    QrCodeResponse:
      type: object
      required: [trackingCode, qrCodeDataUrl, qrCodeUrl]
      properties:
        trackingCode:
          type: string
        qrCodeDataUrl:
          type: string
        qrCodeUrl:
          type: string

    # Dashboard
...
    ProofReviewInput:
      type: object
      required: [reviewStatus]
      properties:
        reviewStatus:
          type: string
          enum: [approved, rejected]
        notes:
          type: ["string", "null"]

    # Reports
...
    PhotoUploadInput:
      type: object
      required: [photo]
      properties:
        photo:
          type: string
          description: Base64-encoded photo data
    PhotoUploadResponse:
      type: object
      required: [url, takenAt]
      properties:
        url:
          type: string
        takenAt:
          type: string
          format: date-time
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]