# lib/api-zod/src/generated/types/trackingDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PDAM Tiara Material Tracking & Verification System API
 * OpenAPI spec version: 0.1.0
 */
import type { InstallationProof } from './installationProof';
import type { TrackingDetailStatus } from './trackingDetailStatus';
import type { TrackingEvent } from './trackingEvent';

export interface TrackingDetail {
  id: string;
  trackingCode: string;
  /** @nullable */
  materialRequestId?: string | null;
  /** @nullable */
  materialName?: string | null;
  /** @nullable */
  materialCode?: string | null;
  /** @nullable */
  materialUnit?: string | null;
  /** @nullable */
  branchName?: string | null;
  /** @nullable */
  branchId?: string | null;
  qtyIssued: number;
  /** @nullable */
  issuedById?: string | null;
  /** @nullable */
  issuedByName?: string | null;
  issuedAt: Date;
  slaDeadline: Date;
  status: TrackingDetailStatus;
  /** @nullable */
  qrCodeUrl?: string | null;
  /** @nullable */
  riskScore?: number | null;
  /** @nullable */
  hoursRemaining?: number | null;
  isOverdue?: boolean;
  createdAt: Date;
  events: TrackingEvent[];
  proof?: InstallationProof;
}